refactor(blocks): structural improvements
