+++ packages/blocks/src/form/components/PasswordInput.tsx
@@ -3,7 +3,7 @@ import { EyeIcon, EyeSlashIcon } from "@heroicons/react/24/outline";
 import { eventHandler } from "@rafty/shared";
 import { Button, InputField, Suffix, useBoolean } from "@rafty/ui";
 import { useFormContext } from "react-hook-form";
-import { FieldWrapper, InputWrapper } from "../../utils";
+import { FieldWrapper, InputWrapper } from "../../utils/components";
 
 export type PasswordInput = {
   data: {
