+++ packages/blocks/src/form/index.ts
@@ -3,7 +3,7 @@ import { BsTextareaT } from "react-icons/bs";
 import { BsBraces } from "react-icons/bs";
 import { LuTextCursorInput } from "react-icons/lu";
 import { MdLink, MdOutlineKey, MdOutlineMailOutline } from "react-icons/md";
-import type { FieldWrapperProps, InputWrapperProps } from "../utils";
+import type { FieldWrapperProps, InputWrapperProps } from "../utils/components";
 import {
   Canvas,
   NumberInput,
