--- packages/blocks/src/form/settings/utils/index.ts
@@ -1,4 +1,0 @@
-export { FieldWrapperSettings } from "./FieldWrapperSettings";
-export { CommonInputSettings } from "./CommonInputSettings";
-export { SettingInput } from "./SettingInput";
-export { SettingsPanelWrapper } from "./SettingsPanelWrapper";
