+++ packages/blocks/src/page/components/Page.tsx
@@ -2,7 +2,7 @@ export type Page = { id: string };
 
 export function Page() {
   return (
-    <div className="bg-secondary-50 h-[1080px] w-[1920px] border">
+    <div className="bg-secondary-50 dark:bg-secondary-950 dark:border-secondary-800 h-[1080px] w-[1920px] border">
       Page Canvas
     </div>
   );
