+++ packages/blocks/src/page/components/index.ts
@@ -1,5 +1,1 @@
 export { Page } from "./Page";
-export { NumberInput } from "./NumberInput";
-export { PasswordInput } from "./PasswordInput";
-export { StringInput } from "./StringInput";
-export { Textarea } from "./Textarea";
