--- packages/blocks/src/page/settings/CanvasSettings.tsx
@@ -1,11 +1,0 @@
-import { Label } from "@rafty/ui";
-import { SettingInput, SettingsPanelWrapper } from "./utils";
-
-export function CanvasSettings() {
-  return (
-    <SettingsPanelWrapper>
-      <Label>Title</Label>
-      <SettingInput name="title" />
-    </SettingsPanelWrapper>
-  );
-}
