--- packages/blocks/src/page/settings/NumberInputSettings.tsx
@@ -1,14 +1,0 @@
-import {
-  CommonInputSettings,
-  FieldWrapperSettings,
-  SettingsPanelWrapper,
-} from "./utils";
-
-export function NumberInputSettings() {
-  return (
-    <SettingsPanelWrapper>
-      <FieldWrapperSettings />
-      <CommonInputSettings />
-    </SettingsPanelWrapper>
-  );
-}
