--- packages/blocks/src/page/settings/PasswordInputSettings.tsx
@@ -1,14 +1,0 @@
-import {
-  CommonInputSettings,
-  FieldWrapperSettings,
-  SettingsPanelWrapper,
-} from "./utils";
-
-export function PasswordInputSettings() {
-  return (
-    <SettingsPanelWrapper>
-      <FieldWrapperSettings />
-      <CommonInputSettings />
-    </SettingsPanelWrapper>
-  );
-}
