--- packages/blocks/src/page/settings/StringInputSettings.tsx
@@ -1,14 +1,0 @@
-import {
-  CommonInputSettings,
-  FieldWrapperSettings,
-  SettingsPanelWrapper,
-} from "./utils";
-
-export function StringInputSettings() {
-  return (
-    <SettingsPanelWrapper>
-      <FieldWrapperSettings />
-      <CommonInputSettings />
-    </SettingsPanelWrapper>
-  );
-}
