--- packages/blocks/src/page/settings/TextareaSettings.tsx
@@ -1,14 +1,0 @@
-import {
-  CommonInputSettings,
-  FieldWrapperSettings,
-  SettingsPanelWrapper,
-} from "./utils";
-
-export function TextareaSettings() {
-  return (
-    <SettingsPanelWrapper>
-      <FieldWrapperSettings />
-      <CommonInputSettings />
-    </SettingsPanelWrapper>
-  );
-}
