+++ packages/blocks/src/page/settings/index.ts
@@ -1,5 +1,1 @@
-export { CanvasSettings } from "./CanvasSettings";
-export { NumberInputSettings } from "./NumberInputSettings";
-export { PasswordInputSettings } from "./PasswordInputSettings";
-export { StringInputSettings } from "./StringInputSettings";
-export { TextareaSettings } from "./TextareaSettings";
+export { PageSettings } from "./PageSettings";
