+++ packages/blocks/src/utils/components/index.ts
@@ -1,0 +1,5 @@
+export { FieldWrapper, type FieldWrapperProps } from "./FieldWrapper";
+export {
+  InputWrapper,
+  type InputWrapperProps,
+} from "../components/InputWrapper";
