--- packages/blocks/src/utils/index.ts
@@ -1,2 +1,0 @@
-export { FieldWrapper, type FieldWrapperProps } from "./FieldWrapper";
-export { InputWrapper, type InputWrapperProps } from "./InputWrapper";
